requirements.txt
flask>=3.0
openai>=1.30
pypdf>=4.0
python-dotenv>=1.0
openpyxl>=3.1
python-docx>=1.1
